requirements.txt
discord.py==2.4.0
google-generativeai==0.8.3